Refactor prm client to be more generic
