+++ libs/clients/src/lib/index.ts
@@ -1,6 +1,6 @@
 import { UserClient } from './user.client';
 
-export * from './prm';
+export * from './prm.client';
 export { BaseClient } from './base.client';
 export { UserClient } from './user.client';
