--- libs/clients/src/lib/prm/index.ts
@@ -1,1 +1,0 @@
-export { PrmClient } from './prm.client';
