--- libs/clients/src/lib/prm/prm.client.test.ts
@@ -1,10 +1,0 @@
-import { PrmClient } from './prm.client';
-
-describe('PrmClient', () => {
-  it('should create an instance', () => {
-    const client = new PrmClient(undefined, {
-      baseURL: 'http://localhost:3000',
-    });
-    expect(client).toBeTruthy();
-  });
-});
